+++ cloud/authentication.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
 )
 
@@ -74,16 +73,15 @@ func (s *AuthenticationService) AcquireSessionCookie(ctx context.Context, userna
 
 	session := new(Session)
 	resp, err := s.client.Do(req, session)
-
-	if resp != nil {
-		session.Cookies = resp.Cookies()
-	}
-
 	if err != nil {
-		return false, fmt.Errorf("auth at Jira instance failed (HTTP(S) request). %s", err)
+		return false, fmt.Errorf("auth at Jira instance failed (HTTP(S) request). %w", err)
 	}
+
 	if resp != nil && resp.StatusCode != 200 {
 		return false, fmt.Errorf("auth at Jira instance failed (HTTP(S) request). Status code: %d", resp.StatusCode)
+	}
+	if resp != nil {
+		session.Cookies = resp.Cookies()
 	}
 
 	s.client.session = session
@@ -128,12 +126,12 @@ func (s *AuthenticationService) Logout(ctx context.Context) error {
 	apiEndpoint := "rest/auth/1/session"
 	req, err := s.client.NewRequest(ctx, http.MethodDelete, apiEndpoint, nil)
 	if err != nil {
-		return fmt.Errorf("creating the request to log the user out failed : %s", err)
+		return fmt.Errorf("creating the request to log the user out failed : %w", err)
 	}
 
 	resp, err := s.client.Do(req, nil)
 	if err != nil {
-		return fmt.Errorf("error sending the logout request: %s", err)
+		return fmt.Errorf("error sending the logout request: %w", err)
 	}
 	defer resp.Body.Close()
 	if resp.StatusCode != 204 {
@@ -161,27 +159,22 @@ func (s *AuthenticationService) GetCurrentUser(ctx context.Context) (*Session, e
 	apiEndpoint := "rest/auth/1/session"
 	req, err := s.client.NewRequest(ctx, http.MethodGet, apiEndpoint, nil)
 	if err != nil {
-		return nil, fmt.Errorf("could not create request for getting user info : %s", err)
+		return nil, fmt.Errorf("could not create request for getting user info: %w", err)
 	}
 
 	resp, err := s.client.Do(req, nil)
 	if err != nil {
-		return nil, fmt.Errorf("error sending request to get user info : %s", err)
+		return nil, fmt.Errorf("error sending request to get user info: %w", err)
 	}
 	defer resp.Body.Close()
 	if resp.StatusCode != 200 {
 		return nil, fmt.Errorf("getting user info failed with status : %d", resp.StatusCode)
 	}
+
 	ret := new(Session)
-	data, err := io.ReadAll(resp.Body)
+	err = json.NewDecoder(resp.Body).Decode(&ret)
 	if err != nil {
-		return nil, fmt.Errorf("couldn't read body from the response : %s", err)
-	}
-
-	err = json.Unmarshal(data, &ret)
-
-	if err != nil {
-		return nil, fmt.Errorf("could not unmarshall received user info : %s", err)
+		return nil, err
 	}
 
 	return ret, nil